+++ Example/MultiSegmentDemo/AppDelegate.swift
@@ -1,0 +1,19 @@
+//
+//  AppDelegate.swift
+//  MultiSegmentDemo
+//
+//  Created by Yonat Sharon on 19/08/2019.
+//  Copyright © 2019 Yonat Sharon. All rights reserved.
+//
+
+import UIKit
+
+@UIApplicationMain
+class AppDelegate: UIResponder, UIApplicationDelegate {
+    var window: UIWindow?
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        // Override point for customization after application launch.
+        return true
+    }
+}